fix stale virus count bug
Players have a "virus placement" mode in which the virus count is
counting up. We consider that mode finished when the virus count reaches
the current level's maximum number of viruses. But this is not quite
right, because we sometimes will see the virus count carried over from a
previous game and not yet reset to 0.

This commit fixes this behavior: virus placement mode now ends after we
have seen a virus count below the max and then see a virus count equal
to the max.